Add AlternateSecondView so that behaviour can be confirmed between two different types of second view.
When push to secondView, if secondView's frame is larger than the frame of firstView, then firstView's frame gets adjusted to the size of secondView's frame.
If secondView's frame is smaller than the frame of firstView, then secondView's frame gets adjusted to the size of firstView's Frame.
Happens between viewDidLoad and viewDidAppear of the next view to be shown.